Fix SCRUM-95: jotai 및 storage 로직 제거 및 login 및 instance 관련 리팩토링
- cookie로 token 관리 시 instance는 api 로직을 분리할 필요가 없기 때문에 1개로 병합
