+++ app/(auth)/login/page.tsx
@@ -1,49 +1,21 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
-/* eslint-disable react/button-has-type */
-// app/login/page.tsx
-
 'use client';
 
-import {useCallback, useLayoutEffect, useState} from 'react';
+import {useState} from 'react';
 import {useForm} from 'react-hook-form';
 
 import {zodResolver} from '@hookform/resolvers/zod';
 import clsx from 'clsx';
-import {useAtom} from 'jotai';
-import _ from 'lodash';
 import {Eye, EyeOff} from 'lucide-react';
 import {useRouter} from 'next/navigation';
 import {z} from 'zod';
 
-import {login, setCookie} from '@/app/apis/user/userApi';
+import {login} from '@/app/apis/user/userApi';
 import useClearAuth from '@/app/hooks/useAuth';
-import {tokenWithStorageAtom} from '@/app/store/atoms/userAtoms';
+import {useDebounce} from '@/app/hooks/useForm';
+import {TLoginInputs} from '@/app/types/auth.types';
+import {LoginSchema} from '@/app/utils/schema';
 
-// zod schema
-// - zod를 이용해 form의 validation schema 정의
-const LoginSchema = z.object({
-  email: z.string().trim().min(1, '이메일을 입력해주세요').email({
-    message: '유효한 이메일 주소를 입력하세요',
-  }),
-  password: z.string().trim().min(1, '비밀번호를 입력해주세요').min(8, {
-    message: '비밀번호는 8자 이상이어야 합니다',
-  }),
-});
-
-type TLoginInputs = z.infer<typeof LoginSchema>;
-
-// Login Page Component
 export default function LoginPage() {
-  // with auth
-  // - 페이지 들어올 때 관련 기능을 초기화
-  const [, setToken] = useAtom(tokenWithStorageAtom);
-  const {clearAuth} = useClearAuth();
-  useLayoutEffect(() => {
-    clearAuth(); // cookie check를 못하기 때문에 무조건 실행
-  }, []);
-
-  // with form
-  // - form state 관리 + validation 시점을 정의 + zod schema를 통한 validation
   const {
     register,
     trigger,
@@ -62,68 +34,33 @@ export default function LoginPage() {
     password: '',
   });
   const [showPassword, setShowPassword] = useState(false);
-
-  // with navigation
+  const {debounceEmailValidate} = useDebounce();
   const router = useRouter();
 
   // Event handlers
   // Form Event
   // Todo: api 처리 로직에 에러 처리 적용하기
   const onSubmit = async (data: TLoginInputs) => {
-    const loginRes = await login(data.email, data.password);
-    if (loginRes === undefined) {
+    try {
+      await login(data.email, data.password);
+      setServerErrorMessage({email: '', password: ''});
+      router.push('/');
+    } catch (error) {
+      useClearAuth();
       alert('로그인 과정에 문제가 발생했습니다.');
-      return;
-    }
-
-    // 로그인 성공
-    // - 1. 토큰 셋팅
-    // - 2. cookie 셋팅
-    // - 3. 메세지 초기화
-    // - 4. router push
-    if ('token' in loginRes) {
-      setToken(loginRes.token); // jotai + sessionStorage에 token 저장
-      const cookieRes = await setCookie();
-      if (cookieRes.status === 200) {
-        setServerErrorMessage({email: '', password: ''});
-        router.push('/');
-      } else {
-        clearAuth();
-        alert('로그인 과정에 문제가 발생했습니다.');
-      }
-    }
-    // 로그인 실패
-    // - 서버에서 보내준 에러를 표시
-    else if (loginRes.code === 'USER_NOT_FOUND' || loginRes.code === 'VALIDATION_ERROR') {
-      setServerErrorMessage({
-        email: loginRes.message,
-        password: '',
-      });
-    } else if (loginRes.code === 'INVALID_CREDENTIALS') {
-      setServerErrorMessage({
-        email: '',
-        password: loginRes.message,
-      });
     }
   };
 
   // Change Event
-  const debounceEmailValidate = useCallback(
-    _.debounce(async (field: keyof TLoginInputs) => {
-      await trigger(field);
-    }, 1000),
-    [],
-  );
-
   const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     const {name} = e.target;
     if (name !== 'email' && name !== 'password') {
       return;
     }
-    debounceEmailValidate(name as keyof TLoginInputs);
+    debounceEmailValidate(name, trigger);
   };
 
-  // render
+  // Render
   return (
     <div className="max-w-[340px] sm:max-w-[600px] xl:max-w-[510px] px-4 py-8 sm:px-14 sm:py-8 flex items-center justify-center flex-col bg-white rounded-3xl w-full">
       <h1 className="text-2xl font-bold text-center mb-8">로그인</h1>
