--- app/store/atoms/userAtoms.ts
@@ -1,25 +1,0 @@
-import {atom} from 'jotai';
-import {atomWithStorage, createJSONStorage} from 'jotai/utils';
-
-import {AUTH_TOKEN} from '@/app/constants/auth';
-
-import {User} from '../types/user.types';
-
-// userInfoAtom
-// - 유저 정보 저자용 atom
-const userInfoAtom = atom<User | null>(null);
-userInfoAtom.debugLabel = 'userInfo';
-
-// tokenWithStorageAtom
-// - token 저장용 atom
-// - sessionStorage 이용
-const sessionStorageProvider = createJSONStorage<string | null>(() => sessionStorage);
-
-const tokenWithStorageAtom = atomWithStorage<string | null>(
-  AUTH_TOKEN,
-  null,
-  sessionStorageProvider,
-);
-tokenWithStorageAtom.debugLabel = 'token with storage';
-
-export {tokenWithStorageAtom, userInfoAtom};
