+++ app/types/auth.types.ts
@@ -1,0 +1,6 @@
+import {z} from 'zod';
+
+import {LoginSchema} from '../utils/schema';
+
+// login types
+export type TLoginInputs = z.infer<typeof LoginSchema>;
